requirements.txt
aioredis==2.0.0
tornado==6.1
python-socketio==5.4.0